requirements.txt
transformers
peft
evaluate
datasets
tokenizers
numpy
pandas
nltk
sentencepiece
protobuf
torchvision
torch
tensorboardX
ninja
huggingface-hub
deepspeed
ujson
pillow

